requirements/base.txt
shimoku-api-python==0.7.1
pandas==1.4.3
requests==2.28.1
python-dotenv==0.20.0
shimoku-components-catalog==0.1.6
